Fix depth buffer #2

+++ lib/Renderer/include/Buffer.h
@@ -24,6 +24,6 @@ public:
 
 	unsigned int* color;
 
-	unsigned int* depth;
+	float* depth;
 };
